refactor(landing): 컨테이너 루트의 동작하지 않는 flex=1 제거 (리뷰 #4 #7)
세로 높이는 데모 안쪽 행의 h=[auto,500px] 하나가 결정하므로 그 위 체인
(DemoTabs 루트/tabpanel, Trans/MathTrans 루트)의 flex=1 은 전부 no-op 이었다.
